# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KYC Panda - x402 KYC Gate API",
    "description": "KYC gate extension for x402 payment protocol. Uses Sign-In-With-X (SIWX/CAIP-122) for wallet-based authentication and Didit for identity verification. Server agents use the bouncer library to gate their endpoints. Client agents interact with this API to onboard and prove KYC status.",
    "version": "1.0.0",
    "contact": {
      "name": "KYC Panda",
      "url": "https://kyc-panda.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://kyc-panda.vercel.app",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/verify/{address}": {
      "get": {
        "operationId": "checkKYCStatus",
        "summary": "Check if a wallet address has completed KYC",
        "description": "Returns the KYC verification status for a given wallet address. This is called by the bouncer hook on every protected request, and can also be called directly by agents to check their status before attempting a gated endpoint.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "EVM wallet address (checksummed or lowercase)",
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "KYC status for the address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                },
                "examples": {
                  "verified": {
                    "summary": "Wallet is KYC verified",
                    "value": { "verified": true }
                  },
                  "notFound": {
                    "summary": "Wallet has never completed KYC",
                    "value": { "verified": false, "reason": "KYC_NOT_FOUND" }
                  },
                  "expired": {
                    "summary": "KYC has expired (older than 1 year)",
                    "value": { "verified": false, "reason": "KYC_EXPIRED" }
                  },
                  "declined": {
                    "summary": "KYC was rejected",
                    "value": { "verified": false, "reason": "KYC_DECLINED" }
                  },
                  "pendingReview": {
                    "summary": "KYC submitted but under manual review",
                    "value": { "verified": false, "reason": "KYC_PENDING_REVIEW" }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (60 requests/minute per IP)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/onboard": {
      "post": {
        "operationId": "startKYCOnboarding",
        "summary": "Start KYC verification for a wallet",
        "description": "Submit a signed SIWX message to begin KYC onboarding. The server verifies your wallet ownership via the SIWX signature, creates a Didit verification session, and returns a URL where the human wallet owner completes identity verification. Once approved, the wallet can pass KYC-gated x402 endpoints.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardRequest"
              },
              "example": {
                "siwxHeader": "eyJkb21haW4iOiJreWMtcGFuZGEudmVyY2VsLmFwcCIsImFkZHJlc3MiOiIweDg1N2IwNjUxOUU5MWUzQTU0NTM4NzkxYkRiYjBFMjIzNzNlMzZiNjYiLCJ1cmkiOiJodHRwczovL2t5Yy1wYW5kYS52ZXJjZWwuYXBwL2FwaS9vbmJvYXJkIiwidmVyc2lvbiI6IjEiLCJjaGFpbklkIjoiZWlwMTU1Ojg0NTMyIiwidHlwZSI6ImVpcDE5MSIsIm5vbmNlIjoiYTFiMmMzZDRlNWY2Nzg5MGExYjJjM2Q0ZTVmNjc4OTAiLCJpc3N1ZWRBdCI6IjIwMjYtMDMtMTBUMTA6MzA6MDAuMDAwWiIsInNpZ25hdHVyZSI6IjB4Li4uIn0="
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "KYC session created. Redirect the human to the verification URL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardResponse"
                },
                "example": {
                  "onboardingId": "550e8400-e29b-41d4-a716-446655440000",
                  "verificationUrl": "https://verify.didit.me/session/abc123"
                }
              }
            }
          },
          "400": {
            "description": "Invalid SIWX message, domain mismatch, expired message, or nonce reuse",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "missingSiwx": {
                    "summary": "Missing required fields",
                    "value": { "error": "Missing siwxMessage or signature" }
                  },
                  "domainMismatch": {
                    "summary": "Domain doesn't match server",
                    "value": { "error": "Domain mismatch" }
                  },
                  "nonceReused": {
                    "summary": "Replay attack detected",
                    "value": { "error": "NONCE_REUSED", "message": "Replay attack detected — nonce already used" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Signature verification failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "INVALID_SIGNATURE", "message": "SIWX signature verification failed" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (10 requests/minute per IP)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "502": {
            "description": "Didit verification service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "503": {
            "description": "KYC service not configured",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "operationId": "healthCheck",
        "summary": "Check service health",
        "description": "Returns service health status and database connectivity.",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["ok"] },
                    "db": { "type": "string", "enum": ["connected"] }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service unhealthy (database disconnected)"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "VerifyResponse": {
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean",
            "description": "true if wallet has valid KYC approval"
          },
          "reason": {
            "type": "string",
            "enum": ["KYC_NOT_FOUND", "KYC_EXPIRED", "KYC_DECLINED", "KYC_PENDING_REVIEW"],
            "description": "Reason for verification failure (only present when verified is false)"
          }
        },
        "required": ["verified"]
      },
      "OnboardRequest": {
        "type": "object",
        "properties": {
          "siwxHeader": {
            "type": "string",
            "description": "Base64-encoded SIWX payload JSON. Same format as the SIGN-IN-WITH-X header in x402 requests. Must include domain matching kyc-panda.vercel.app, a fresh nonce, issuedAt within 5 minutes, and a valid wallet signature. Create with createSIWxPayload() + encodeSIWxHeader() from @x402/extensions/sign-in-with-x."
          }
        },
        "required": ["siwxHeader"]
      },
      "OnboardResponse": {
        "type": "object",
        "properties": {
          "onboardingId": {
            "type": "string",
            "format": "uuid",
            "description": "Internal tracking ID for this onboarding session"
          },
          "verificationUrl": {
            "type": "string",
            "format": "uri",
            "description": "Didit hosted verification URL. The human wallet owner should open this to complete KYC."
          }
        },
        "required": ["onboardingId", "verificationUrl"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code or message"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error detail (optional)"
          }
        },
        "required": ["error"]
      },
      "KYCGateExtension": {
        "type": "object",
        "description": "The kyc-gate extension object that appears in x402 402 responses alongside sign-in-with-x. Tells agents that KYC verification is required.",
        "properties": {
          "required": {
            "type": "boolean",
            "description": "Always true — KYC is mandatory for this endpoint"
          },
          "onboardingUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to POST SIWX message + signature to begin KYC"
          },
          "provider": {
            "type": "string",
            "description": "KYC provider name (e.g. didit)"
          }
        },
        "required": ["required", "onboardingUrl", "provider"]
      },
      "SIWxChallenge": {
        "type": "object",
        "description": "The sign-in-with-x extension object in a 402 response. Contains the SIWX challenge that clients must sign.",
        "properties": {
          "info": {
            "type": "object",
            "description": "Challenge metadata per CAIP-122",
            "properties": {
              "domain": { "type": "string" },
              "uri": { "type": "string", "format": "uri" },
              "version": { "type": "string", "enum": ["1"] },
              "nonce": { "type": "string", "minLength": 32, "maxLength": 32 },
              "issuedAt": { "type": "string", "format": "date-time" },
              "expirationTime": { "type": "string", "format": "date-time" },
              "statement": { "type": "string" },
              "resources": { "type": "array", "items": { "type": "string", "format": "uri" } }
            },
            "required": ["domain", "uri", "version", "nonce", "issuedAt"]
          },
          "supportedChains": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "chainId": { "type": "string", "description": "CAIP-2 chain ID (e.g. eip155:8453)" },
                "type": { "type": "string", "description": "Signature type (eip191 for EVM)" }
              },
              "required": ["chainId", "type"]
            }
          },
          "schema": {
            "type": "object",
            "description": "JSON Schema for the client proof payload"
          }
        },
        "required": ["info", "supportedChains", "schema"]
      }
    }
  }
}